+++ CesiumGeometry/include/CesiumGeometry/IntersectionTests.h
@@ -29,6 +29,44 @@ public:
   rayPlane(const Ray& ray, const Plane& plane) noexcept;
 
   /**
+   * @brief Tests if a ray hits a triangle and returns the hit point, if any.
+   *
+   * @param ray The ray.
+   * @param p0 The first vertex of the triangle.
+   * @param p1 The second vertex of the triangle.
+   * @param p2 The third vertex of the triangle.
+   * @param cullBackFaces Whether to cull back faces or not.
+   * @return The hit point, if any.
+   */
+  static std::optional<glm::dvec3> rayTriangle(
+      const Ray& ray,
+      const glm::dvec3& p0,
+      const glm::dvec3& p1,
+      const glm::dvec3& p2,
+      bool cullBackFaces = false);
+
+  /**
+   * @brief Tests if a ray hits a triangle and outputs the distance to the hit
+   * point, if any.
+   *
+   * @param ray The ray.
+   * @param p0 The first vertex of the triangle.
+   * @param p1 The second vertex of the triangle.
+   * @param p2 The third vertex of the triangle.
+   * @param cullBackFaces Whether to cull back faces or not.
+   * @param[out] t The distance from the ray origin to the intersection point,
+   * if any.
+   * @return Whether the ray intersects the triangle.
+   */
+  static bool rayTriangleParametric(
+      const Ray& ray,
+      const glm::dvec3& p0,
+      const glm::dvec3& p1,
+      const glm::dvec3& p2,
+      double& t,
+      bool cullBackFaces = false);
+
+  /**
    * @brief Determines whether the point is completely inside the triangle.
    *
    * @param point The point to check.